bump minimum supported Map::Tube version
... to the smallest version of Map::Tube for which
tests seem to pass at least on my perl version.

Map::Tube v3.39 fixed the open() line for the JSON file,
which otherwise made the test suite for Map::Tube::JSON
die with a "ERROR: Mapformed Map Data ...".
Earlier versions of Map::Tube otherwise instead required
to provide an "xml" as well.

Thus, v3.39 is the first version this module can truly
"support", as it is the first one in which the test
suite does not die.